src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic-settings==2.9.1
aiofiles==24.1.0
regex==2024.11.6
langchain==0.3.25
PyMuPDF==1.26.0
langchain-community
